Выносит коннект к стору из хока управления плеером в родительский компонент
