+++ src/components/app/app.jsx
@@ -1,5 +1,5 @@
 import React, {PureComponent} from "react";
-import {Switch, Route, Router, withRouter} from "react-router-dom";
+import {Switch, Route, Router} from "react-router-dom";
 import history from "../../history.js";
 
 import MainScreen from "../main-screen/main-screen.jsx";
@@ -12,13 +12,11 @@ import VideoPlayerPage from "../video-player-page/video-player-page.jsx";
 import {MovieInfoTabs, DEFAULT_GENRE, AppRoute} from "../../consts.js";
 import withActiveTab from "../../hocs/with-active-tab.jsx";
 import withPrivateRoute from "../../hocs/with-private-route.jsx";
-import withVideoPlayerControls from "../../hocs/with-video-player-controls.jsx";
 
 const DEFAULT_MOVIE_INFO_TAB = MovieInfoTabs.OVERVIEW;
 
 const MovieInfoWrapped = withActiveTab(MovieInfo, DEFAULT_MOVIE_INFO_TAB);
 const MainScreenWrapped = withActiveTab(MainScreen, DEFAULT_GENRE);
-const VideoPlayerPageWrapped = withVideoPlayerControls(VideoPlayerPage);
 
 // gперенести withrouter в компоненты
 
@@ -36,7 +34,7 @@ class App extends PureComponent {
           <Route exact path={AppRoute.LOGIN} component={SignInPage} />
           <Route exact path={AppRoute.MY_LIST} component={withPrivateRoute(MyListPage)} />
           <Route exact path={`${AppRoute.FILMS}/:id${AppRoute.REVIEW}`} component={withPrivateRoute(AddReviewPage)} />
-          <Route exact path={`${AppRoute.PLAYER}/:id`} component={withRouter(VideoPlayerPageWrapped)} />
+          <Route exact path={`${AppRoute.PLAYER}/:id`} component={VideoPlayerPage} />
         </Switch>
       </Router>
     );
