+++ src/components/video-player-page/video-player-page.jsx
@@ -1,66 +1,40 @@
 import React from "react";
 import PropTypes from "prop-types";
+import {connect} from "react-redux";
+import {withRouter} from "react-router-dom";
+import history from "../../history.js";
 
-import VideoPlayer from "../video-player/video-player.jsx";
-import VideoPlayerTimeline from "../video-player-timeline/video-player-timeline.jsx";
-import VideoPlayerPlayButton from "../video-player-play-button/video-player-play-button.jsx";
-import VideoPlayerPauseButton from "../video-player-pause-button/video-player-pause-button.jsx";
-import VideoPlayerFullscreenButton from "../video-player-fullscreen-button/video-player-fullscreen-button.jsx";
-import VideoPlayerExitButton from "../video-player-exit-button/video-player-exit-button.jsx";
+import VideoPlayerContainer from "../video-player-container/video-player-container.jsx";
 
+import withVideoPlayerControls from "../../hocs/with-video-player-controls.jsx";
+import {getMovieById} from "../../reducer/data/selectors.js";
+import {AppRoute} from "../../consts.js";
 
-export default function VideoPlayerPage(props) {
-  const {
-    movie, isPlaying, videoRef, progress, remaining,
-    onExitButtonClickHandler,
-    onPauseButtonClickHandler,
-    onPlayButtonClickHandler,
-    onFullscreenButtonClickHandler
-  } = props;
-  const {videoFull, preview, title} = movie;
+const WrappedVideoPlayerContainer = withVideoPlayerControls(VideoPlayerContainer);
+
+function VideoPlayerPage({movie}) {
+
+  const onExitButtonClickHandler = () => {
+    return history.push(`${AppRoute.FILMS}/${movie.id}`);
+  };
 
   return (
-    <div className="player">
-      <VideoPlayer
-        src={videoFull}
-        poster={preview}
-        muted={true}
-        extraClass={`player__video`}
-        videoRef={videoRef}
-        loop={false}
-      />
-
-      <VideoPlayerExitButton onClick={onExitButtonClickHandler} />
-
-      <div className="player__controls">
-        <VideoPlayerTimeline progress={progress} remaining={remaining} />
-        <div className="player__controls-row">
-          {isPlaying
-            ? <VideoPlayerPauseButton onClick={onPauseButtonClickHandler} />
-            : <VideoPlayerPlayButton onClick={onPlayButtonClickHandler} />
-          }
-          <div className="player__name">{title}</div>
-          <VideoPlayerFullscreenButton onClick={onFullscreenButtonClickHandler}/>
-
-        </div>
-      </div>
-    </div>
+    <WrappedVideoPlayerContainer
+      onExitButtonClickHandler={onExitButtonClickHandler}
+      movie={movie}
+    />
   );
 }
 
+const mapStateToProps = (state, props) => ({
+  movie: getMovieById(state, props.match.params.id)
+});
+
 VideoPlayerPage.propTypes = {
   movie: PropTypes.shape({
-    videoFull: PropTypes.string.isRequired,
-    preview: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-  }).isRequired,
-  isPlaying: PropTypes.bool.isRequired,
-  videoRef: PropTypes.object.isRequired,
-  progress: PropTypes.number.isRequired,
-  remaining: PropTypes.number.isRequired,
-  onExitButtonClickHandler: PropTypes.func.isRequired,
-  onPauseButtonClickHandler: PropTypes.func.isRequired,
-  onPlayButtonClickHandler: PropTypes.func.isRequired,
-  onFullscreenButtonClickHandler: PropTypes.func.isRequired,
+    id: PropTypes.string.isRequired,
+  })
 };
 
+export {VideoPlayerPage};
+export default withRouter(connect(mapStateToProps)(VideoPlayerPage));
